move system actor migration to its own file
