+++ actors/migration/nv16/system.go
@@ -1,0 +1,29 @@
+package nv16
+
+import (
+	"context"
+
+	system8 "github.com/filecoin-project/specs-actors/v8/actors/builtin/system"
+	"github.com/ipfs/go-cid"
+	cbor "github.com/ipfs/go-ipld-cbor"
+)
+
+// System Actor migrator
+type systemActorMigrator struct {
+	OutCodeCID   cid.Cid
+	ManifestData cid.Cid
+}
+
+func (m systemActorMigrator) migrateState(ctx context.Context, store cbor.IpldStore, in actorMigrationInput) (*actorMigrationResult, error) {
+	state := system8.State{m.ManifestData}
+	stateHead, err := store.Put(ctx, &state)
+
+	if err != nil {
+		return nil, err
+	}
+
+	return &actorMigrationResult{
+		newCodeCID: m.OutCodeCID,
+		newHead:    stateHead,
+	}, nil
+}
